+++ riverdriver/river_driver_interface.go
@@ -383,6 +383,7 @@ type JobDeleteBeforeParams struct {
 	DiscardedDoDelete           bool
 	DiscardedFinalizedAtHorizon time.Time
 	Max                         int
+	Queues                      []string
 	QueuesExcluded              []string
 	QueuesIncluded              []string
 	Schema                      string
@@ -422,9 +423,10 @@ type JobGetByKindManyParams struct {
 }
 
 type JobGetStuckParams struct {
-	Max          int
-	Schema       string
-	StuckHorizon time.Time
+	Max            int
+	QueuesIncluded []string
+	Schema         string
+	StuckHorizon   time.Time
 }
 
 type JobInsertFastParams struct {
@@ -515,9 +517,10 @@ type JobRetryParams struct {
 }
 
 type JobScheduleParams struct {
-	Max    int
-	Now    *time.Time
-	Schema string
+	Max            int
+	Now            *time.Time
+	QueuesIncluded []string
+	Schema         string
 }
 
 type JobScheduleResult struct {
@@ -689,6 +692,7 @@ type LeaderInsertParams struct {
 	ElectedAt *time.Time
 	ExpiresAt *time.Time
 	LeaderID  string
+	Name      string
 	Now       *time.Time
 	Schema    string
 	TTL       time.Duration
@@ -696,6 +700,7 @@ type LeaderInsertParams struct {
 
 type LeaderElectParams struct {
 	LeaderID string
+	Name     string
 	Now      *time.Time
 	Schema   string
 	TTL      time.Duration
@@ -704,6 +709,7 @@ type LeaderElectParams struct {
 type LeaderReelectParams struct {
 	ElectedAt time.Time
 	LeaderID  string
+	Name      string
 	Now       *time.Time
 	Schema    string
 	TTL       time.Duration
@@ -793,6 +799,7 @@ type QueueCreateOrSetUpdatedAtParams struct {
 
 type QueueDeleteExpiredParams struct {
 	Max              int
+	QueuesIncluded   []string
 	Schema           string
 	UpdatedAtHorizon time.Time
 }