+++ riverdriver/riverdatabasesql/internal/dbsqlc/river_leader.sql.go
@@ -15,12 +15,14 @@ const leaderAttemptElect = `-- name: LeaderAttemptElect :one
 INSERT INTO /* TEMPLATE: schema */river_leader (
     leader_id,
     elected_at,
-    expires_at
+    expires_at,
+    name
 ) VALUES (
     $1,
     coalesce($2::timestamptz, now()),
     -- @ttl is inserted as as seconds rather than a duration because ` + "`" + `lib/pq` + "`" + ` doesn't support the latter
-    coalesce($2::timestamptz, now()) + make_interval(secs => $3)
+    coalesce($2::timestamptz, now()) + make_interval(secs => $3),
+    $4
 )
 ON CONFLICT (name)
     DO NOTHING
@@ -31,10 +33,16 @@ type LeaderAttemptElectParams struct {
 	LeaderID string
 	Now      *time.Time
 	TTL      float64
+	Name     string
 }
 
 func (q *Queries) LeaderAttemptElect(ctx context.Context, db DBTX, arg *LeaderAttemptElectParams) (*RiverLeader, error) {
-	row := db.QueryRowContext(ctx, leaderAttemptElect, arg.LeaderID, arg.Now, arg.TTL)
+	row := db.QueryRowContext(ctx, leaderAttemptElect,
+		arg.LeaderID,
+		arg.Now,
+		arg.TTL,
+		arg.Name,
+	)
 	var i RiverLeader
 	err := row.Scan(
 		&i.ElectedAt,
@@ -52,6 +60,7 @@ WHERE
     elected_at = $3::timestamptz
     AND expires_at >= coalesce($1::timestamptz, now())
     AND leader_id = $4
+    AND name = $5
 RETURNING elected_at, expires_at, leader_id, name
 `
 
@@ -60,6 +69,7 @@ type LeaderAttemptReelectParams struct {
 	TTL       float64
 	ElectedAt time.Time
 	LeaderID  string
+	Name      string
 }
 
 func (q *Queries) LeaderAttemptReelect(ctx context.Context, db DBTX, arg *LeaderAttemptReelectParams) (*RiverLeader, error) {
@@ -68,6 +78,7 @@ func (q *Queries) LeaderAttemptReelect(ctx context.Context, db DBTX, arg *Leader
 		arg.TTL,
 		arg.ElectedAt,
 		arg.LeaderID,
+		arg.Name,
 	)
 	var i RiverLeader
 	err := row.Scan(
@@ -113,11 +124,13 @@ const leaderInsert = `-- name: LeaderInsert :one
 INSERT INTO /* TEMPLATE: schema */river_leader(
     elected_at,
     expires_at,
-    leader_id
+    leader_id,
+    name
 ) VALUES (
     coalesce($1::timestamptz, coalesce($2::timestamptz, now())),
     coalesce($3::timestamptz, coalesce($2::timestamptz, now()) + make_interval(secs => $4)),
-    $5
+    $5,
+    $6
 ) RETURNING elected_at, expires_at, leader_id, name
 `
 
@@ -127,6 +140,7 @@ type LeaderInsertParams struct {
 	ExpiresAt *time.Time
 	TTL       float64
 	LeaderID  string
+	Name      string
 }
 
 func (q *Queries) LeaderInsert(ctx context.Context, db DBTX, arg *LeaderInsertParams) (*RiverLeader, error) {
@@ -136,6 +150,7 @@ func (q *Queries) LeaderInsert(ctx context.Context, db DBTX, arg *LeaderInsertPa
 		arg.ExpiresAt,
 		arg.TTL,
 		arg.LeaderID,
+		arg.Name,
 	)
 	var i RiverLeader
 	err := row.Scan(