+++ riverdriver/riverpgxv5/river_pgx_v5_driver.go
@@ -275,7 +275,7 @@ func (e *Executor) JobDeleteBefore(ctx context.Context, params *riverdriver.JobD
 		DiscardedFinalizedAtHorizon: params.DiscardedFinalizedAtHorizon,
 		Max:                         int64(params.Max),
 		QueuesExcluded:              params.QueuesExcluded,
-		QueuesIncluded:              params.QueuesIncluded,
+		QueuesIncluded:              sliceutil.NilIfEmpty(params.QueuesIncluded),
 	})
 	if err != nil {
 		return 0, interpretError(err)
@@ -336,8 +336,9 @@ func (e *Executor) JobGetByKindMany(ctx context.Context, params *riverdriver.Job
 
 func (e *Executor) JobGetStuck(ctx context.Context, params *riverdriver.JobGetStuckParams) ([]*rivertype.JobRow, error) {
 	jobs, err := dbsqlc.New().JobGetStuck(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.JobGetStuckParams{
-		Max:          int32(min(params.Max, math.MaxInt32)), //nolint:gosec
-		StuckHorizon: params.StuckHorizon,
+		Max:            int32(min(params.Max, math.MaxInt32)), //nolint:gosec
+		QueuesIncluded: sliceutil.NilIfEmpty(params.QueuesIncluded),
+		StuckHorizon:   params.StuckHorizon,
 	})
 	if err != nil {
 		return nil, interpretError(err)
@@ -589,8 +590,9 @@ func (e *Executor) JobRetry(ctx context.Context, params *riverdriver.JobRetryPar
 
 func (e *Executor) JobSchedule(ctx context.Context, params *riverdriver.JobScheduleParams) ([]*riverdriver.JobScheduleResult, error) {
 	scheduleResults, err := dbsqlc.New().JobSchedule(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.JobScheduleParams{
-		Max: int64(params.Max),
-		Now: params.Now,
+		Max:            int64(params.Max),
+		Now:            params.Now,
+		QueuesIncluded: sliceutil.NilIfEmpty(params.QueuesIncluded),
 	})
 	if err != nil {
 		return nil, interpretError(err)
@@ -704,6 +706,7 @@ func (e *Executor) JobUpdateFull(ctx context.Context, params *riverdriver.JobUpd
 func (e *Executor) LeaderAttemptElect(ctx context.Context, params *riverdriver.LeaderElectParams) (*riverdriver.Leader, error) {
 	leader, err := dbsqlc.New().LeaderAttemptElect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptElectParams{
 		LeaderID: params.LeaderID,
+		Name:     params.Name,
 		Now:      params.Now,
 		TTL:      params.TTL.Seconds(),
 	})
@@ -717,6 +720,7 @@ func (e *Executor) LeaderAttemptReelect(ctx context.Context, params *riverdriver
 	leader, err := dbsqlc.New().LeaderAttemptReelect(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.LeaderAttemptReelectParams{
 		ElectedAt: params.ElectedAt,
 		LeaderID:  params.LeaderID,
+		Name:      params.Name,
 		Now:       params.Now,
 		TTL:       params.TTL.Seconds(),
 	})
@@ -747,6 +751,7 @@ func (e *Executor) LeaderInsert(ctx context.Context, params *riverdriver.LeaderI
 		ElectedAt: params.ElectedAt,
 		ExpiresAt: params.ExpiresAt,
 		LeaderID:  params.LeaderID,
+		Name:      params.Name,
 		Now:       params.Now,
 		TTL:       params.TTL.Seconds(),
 	})
@@ -874,11 +879,13 @@ func (e *Executor) QueueCreateOrSetUpdatedAt(ctx context.Context, params *riverd
 func (e *Executor) QueueDeleteExpired(ctx context.Context, params *riverdriver.QueueDeleteExpiredParams) ([]string, error) {
 	queues, err := dbsqlc.New().QueueDeleteExpired(schemaTemplateParam(ctx, params.Schema), e.dbtx, &dbsqlc.QueueDeleteExpiredParams{
 		Max:              int64(params.Max),
+		QueuesIncluded:   sliceutil.NilIfEmpty(params.QueuesIncluded),
 		UpdatedAtHorizon: params.UpdatedAtHorizon,
 	})
 	if err != nil {
 		return nil, interpretError(err)
 	}
+
 	queueNames := make([]string, len(queues))
 	for i, q := range queues {
 		queueNames[i] = q.Name